Web/internal/bugfixes (#237)
* Fix page height height for assessment page

* Potential fix for issue on prod (trying to load applicants from an empty cache)

* do NOT persist firestoreReducer, way too much data in DB to persist without exceeding quota

* bump dependencies

* Fix indeterminate checkbox hover state

* Optimize rerendering of <ApplicantList /> and < GenericApplicantContainer />

Only re-render when needed. This prevents the browser from hanging indefinitely.

However, this does not prevent shouldComponentUpdate() from being called, right now if you tag all 1600 applicants, it will call shouldComponentUpdate() 6000 times (because the updates from the databases don't come in with one update) which causes the UI to become unresponsive. So there will still be a multi-second delay (but the tab wouldn't crash at the very least).

* make build